+++ cas-server-core/src/test/java/org/jasig/cas/services/JsonServiceRegistryDaoTests.java
@@ -21,7 +21,6 @@ package org.jasig.cas.services;
 import org.apache.commons.io.FileUtils;
 import org.jasig.cas.authentication.principal.CachingPrincipalAttributesRepository;
 import org.jasig.cas.authentication.principal.ShibbolethCompatiblePersistentIdGenerator;
-import org.jasig.cas.services.support.DefaultRegisteredServiceAuthorizationStrategy;
 import org.jasig.cas.services.support.RegisteredServiceRegexAttributeFilter;
 import org.jasig.services.persondir.support.StubPersonAttributeDao;
 import org.jasig.services.persondir.support.merger.ReplacingAttributeAdder;
