+++ cas-server-webapp-support/src/test/java/org/jasig/cas/web/flow/ServiceAuthorizationCheckTests.java
@@ -26,7 +26,7 @@ import org.jasig.cas.services.RegisteredService;
 import org.jasig.cas.services.RegisteredServiceImpl;
 import org.jasig.cas.services.ServicesManager;
 import org.jasig.cas.services.UnauthorizedServiceException;
-import org.jasig.cas.services.support.DefaultRegisteredServiceAuthorizationStrategy;
+import org.jasig.cas.services.DefaultRegisteredServiceAuthorizationStrategy;
 import org.junit.Before;
 import org.junit.Test;
 import org.springframework.webflow.execution.Event;
